[FIX] refactoring after PR
