+++ dahee/app.js
@@ -4,6 +4,7 @@ const express = require('express');
 const cors = require('cors');
 const morgan = require('morgan');
 const routes = require('./routes');
+const appDataSource = require('./models/appDataSource');
 
 const app = express();
 const PORT = process.env.PORT;
@@ -12,6 +13,15 @@ app.use(cors());
 app.use(express.json());
 app.use(morgan('combined'));
 app.use(routes);
+
+appDataSource
+  .initialize()
+  .then(() => {
+    console.log('Data Source has been initialized!');
+  })
+  .catch((err) => {
+    console.error('Error during Data Source initialization', err);
+  });
 
 app.get('/ping', (req, res) => {
   res.status(200).json({ message: 'pong' });
