+++ dahee/controllers/postController.js
@@ -51,7 +51,23 @@ const deletePost = async (req, res) => {
       return res.status(400).json({ message: 'KEY_ERROR' });
     }
     await postService.deletePost(postId);
-    res.status(204).json({ message: 'POST_DELETED' });
+    res.status(204).send();
+  } catch (err) {
+    console.log(err);
+    return res.status(err.statusCode || 500).json({ message: err.message });
+  }
+};
+
+const postsByUser = async (req, res) => {
+  try {
+    const { userId } = req.params;
+    if (!userId) {
+      return res.status(400).json({ message: 'KEY_ERROR' });
+    }
+
+    const data = await postService.postsByUser(userId);
+
+    res.status(200).json({ data });
   } catch (err) {
     console.log(err);
     return res.status(err.statusCode || 500).json({ message: err.message });
@@ -63,4 +79,5 @@ module.exports = {
   getAllPosts,
   updatePost,
   deletePost,
+  postsByUser,
 };
