+++ dahee/controllers/userController.js
@@ -17,23 +17,6 @@ const signUp = async (req, res) => {
   }
 };
 
-const postsByUser = async (req, res) => {
-  try {
-    const { userId } = req.params;
-    if (!userId) {
-      return res.status(400).json({ message: 'KEY_ERROR' });
-    }
-
-    const data = await userService.postsByUser(userId);
-
-    res.status(200).json({ data });
-  } catch (err) {
-    console.log(err);
-    return res.status(err.statusCode || 500).json({ message: err.message });
-  }
-};
-
 module.exports = {
   signUp,
-  postsByUser,
 };
