+++ dahee/models/appDataSource.js
@@ -1,4 +1,3 @@
-require('dotenv').config();
 const { DataSource } = require('typeorm');
 
 const appDataSource = new DataSource({
@@ -10,15 +9,4 @@ const appDataSource = new DataSource({
   database: process.env.DB_DATABASE,
 });
 
-appDataSource
-  .initialize()
-  .then(() => {
-    console.log('Data Source has been initialized!');
-  })
-  .catch((err) => {
-    console.error('Error during Data Source initialization', err);
-  });
-
-module.exports = {
-  appDataSource,
-};
+module.exports = appDataSource;
