+++ dahee/models/userDao.js
@@ -1,8 +1,8 @@
-const dataSource = require('./appDataSource');
+const appDataSource = require('./appDataSource');
 
 const createUser = async (name, email, password, profileImage) => {
   try {
-    return await dataSource.appDataSource.query(
+    return await appDataSource.query(
       `INSERT INTO users(
                   name,
                   email,
@@ -18,37 +18,6 @@ const createUser = async (name, email, password, profileImage) => {
   }
 };
 
-const getUserPosts = async (userId) => {
-  try {
-    return await dataSource.appDataSource.query(
-      `SELECT
-            u.id userId,
-            u.profile_image userProfileImage,
-            pj.postings
-          FROM users u
-          JOIN 
-            (SELECT 
-              user_id, 
-              JSON_ARRAYAGG(
-                JSON_OBJECT(
-                  "postingId", id, 
-                  "postingImageUrl", image_url, 
-                  "postingContent", content
-                  )
-                ) postings
-            FROM posts p
-            GROUP BY id) pj
-          ON u.id = pj.user_id
-          WHERE u.id = ${userId};`
-    );
-  } catch (err) {
-    const error = new Error('INVALID_DATA_INPUT');
-    error.statusCode = 400;
-    throw error;
-  }
-};
-
 module.exports = {
   createUser,
-  getUserPosts,
 };
