+++ dahee/routes/postRouter.js
@@ -7,6 +7,7 @@ try {
   router.get('', postController.getAllPosts);
   router.patch('/:userId/:postId', postController.updatePost);
   router.delete('/:postId', postController.deletePost);
+  router.get('/:userId', postController.postsByUser);
 } catch (err) {
   console.log(err);
 }
