+++ dahee/services/postService.js
@@ -1,19 +1,16 @@
 const postDao = require('../models/postDao');
 
 const newPostUp = async (title, content, userId) => {
-  const postUpload = await postDao.postUpload(title, content, userId);
-  return postUpload;
+  return postDao.postUpload(title, content, userId);
 };
 
 const getAllPosts = async () => {
-  const allPosts = await postDao.allPosts();
-  return allPosts;
+  return postDao.allPosts();
 };
 
 const updatePost = async (userId, postId, title, content) => {
   try {
-    const postChange = await postDao.postChange(userId, postId, title, content);
-    return postChange;
+    return postDao.postChange(userId, postId, title, content);
   } catch (err) {
     console.log(err);
   }
@@ -21,11 +18,14 @@ const updatePost = async (userId, postId, title, content) => {
 
 const deletePost = async (postId) => {
   try {
-    const postDeleting = await postDao.postDeleting(postId);
-    return postDeleting;
+    return postDao.postDeleting(postId);
   } catch (err) {
     console.log(err);
   }
+};
+
+const postsByUser = async (userId) => {
+  return postDao.getUserPosts(userId);
 };
 
 module.exports = {
@@ -33,4 +33,5 @@ module.exports = {
   getAllPosts,
   updatePost,
   deletePost,
+  postsByUser,
 };
