+++ dahee/services/userService.js
@@ -10,27 +10,9 @@ const signUp = async (name, email, password, profileImage) => {
     err.statusCode = 400;
     throw err;
   }
-  const createUser = await userDao.createUser(
-    name,
-    email,
-    password,
-    profileImage
-  );
-  return createUser;
-};
-
-const postsByUser = async (userId) => {
-  if (!userId) {
-    const err = new Error('USERID_IS_NOT_VALID');
-    err.statusCode = 400;
-    throw err;
-  }
-
-  const getUserPosts = await userDao.getUserPosts(userId);
-  return getUserPosts;
+  return userDao.createUser(name, email, password, profileImage);
 };
 
 module.exports = {
   signUp,
-  postsByUser,
 };
